+++ px_install/util.py
@@ -1,11 +1,15 @@
 '''Misc'''
 
+import json
 import os
 import random
 import re
 import string
 import subprocess
 
+import cpuinfo
+import qrcode
+import urllib3
 from pytz import all_timezones
 from tqdm import tqdm
 
@@ -49,10 +53,10 @@ def is_valid_hostname(hostname) -> bool:
 
 
 def is_valid_timezone(timezone: str) -> bool:
-	if timezone in all_timezones:
-		return True
-	else:
-		return False
+    if timezone in all_timezones:
+        return True
+    else:
+        return False
 
 
 def list_of_commands_to_string(command: list):
@@ -76,9 +80,82 @@ def run_commands(commands: list, show_progress: bool = True):
             res = subprocess.run(command_string, shell=True, stdout=subprocess.DEVNULL)
             if res.stderr:
                 print(res.stderr)
+                raise Exception(res.stderr)
     else:
         for command in commands:
             command_string = list_of_commands_to_string(command)
             res = subprocess.run(command_string, shell=True, stdout=subprocess.DEVNULL)
             if res.stderr:
                 print(res.stderr)
+                raise Exception(res.stderr)
+
+
+def convert_size_string(value: str):
+    size = 0
+    if 'T' in value:
+        clean = value[:-1]
+        size = float(clean) * 1000000
+    elif 'G' in value:
+        clean = value[:-1]
+        size = float(clean) * 1000
+    elif 'M' in value:
+        clean = value[:-1]
+        size = float(clean)
+    else:
+        raise ValueError('Could not convert {} to a number.'.format(value))
+    return size
+
+
+def pre_install_environment_check(config):
+    disk = config.disk
+    disk.reload()
+
+    errors = []
+
+    data_partition = disk.get_partition_dev_name(2)
+    data_partition_valid = disk.is_valid_partition(data_partition)
+    if not data_partition_valid:
+        errors.append('Partition {} does not exist')
+
+    swap_file_found = os.path.isfile('/mnt/swapfile')
+    if not swap_file_found:
+        errors.append('Swap file /mnt/swapfile does not exist.')
+
+
+    system_config_found = os.path.isfile('/mnt/etc/system.scm')
+    if not system_config_found:
+        errors.append('System config /mnt/etc/system.scm does not exist.')
+
+
+    if len(errors) > 0:
+        for error in errors:
+            print(error)
+        raise EnvironmentError('Installation pre-requirement(s) check failed.')
+
+
+def print_debug_qr_code(debug: str):
+    qr = qrcode.QRCode(error_correction=qrcode.constants.ERROR_CORRECT_L)
+    qr.add_data(debug)
+    qr.print_ascii(invert=True)
+
+
+def generate_and_print_debug_info(config, version, error=''):
+    error_truncate = (error[:50] + '..') if len(error) > 50 else error
+    debug_output = {
+        'v': version,
+        'c': cpuinfo.get_cpu_info()['brand_raw'],
+        'f': config.firmware,
+        # 'l': config.locale,
+        'd': config.disk.name,
+        'e': error_truncate,
+    }
+    print_debug_qr_code(json.dumps(debug_output))
+
+
+def is_online():
+    try:
+        http = urllib3.PoolManager()
+        http.request('GET', 'http://138.201.123.174', timeout=1)
+        return True
+    except urllib3.exceptions.ProtocolError:
+        return False